requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
numpy>=1.24.0
pypinyin>=0.49.0
edge-tts>=6.1.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
gTTS>=2.4.0
python-dotenv>=1.0.0